fixed issues with querying * changed the line number in the Dictionary to a pointer into the file * added the Query class to support semantic fields * changed the reading of the inverted file to use RandomAccessFile * added the GloSim files as resources

+++ Scobo/src/indexer/Dictionary.java
@@ -329,7 +329,7 @@ public final class Dictionary {
             String term = contents[0];
             int termFrequency = Integer.parseInt(contents[1]);
             int documentFrequency = Integer.parseInt(contents[2]);
-            int pointer = Integer.parseInt(contents[3]);
+            long pointer = Long.parseLong(contents[3]);
             res.dictionary.put(term, new Term(term, termFrequency, documentFrequency, pointer));
         }
